compute: Update quota sets schema
We extended Nova in our env to support quota for instances and will also
support hardware versions for cores and ram. To support this, we extend
the schema used to check update and get requests against quota sets to
include attributes matching by pattern. This allows us to create any
such quota without having to update the test every time.